[Refactor] : 레거시코드 home 모듈로 이동
Git : #1
