--- lib/_testt/modules/auth/register/model/register_model.dart
@@ -1,11 +1,0 @@
-class RegisterRequest {
-  final String email;
-  final String password1;
-  final String password2;
-
-  RegisterRequest({
-    required this.email,
-    required this.password1,
-    required this.password2,
-  });
-}
